Move filter stage into own file
